Retrieve pod metrics only if metrics server is running and Go lint fixes (#2094)
* Retrieve pod metrics only if metrics server is running

Currently we query pod metrics every 30 sec which floods executor logs,
added check which confirms if metrics server is running then only we start
querying pod metrics for identifying CPU utilization.


* Fixed couple of typos and misspells with Go CI


* Remove unnecessary conversions with Go CI

+++ pkg/executor/fscache/metrics.go
@@ -88,7 +88,7 @@ func (fsc *FunctionServiceCache) setFuncAlive(funcname, funcuid string, isAlive 
 
 // ReapTime is the amount of time taken to reap a pod
 func (fsc *FunctionServiceCache) ReapTime(funcName, funcAddress string, time float64) {
-	funcReapTime.WithLabelValues(funcName, funcAddress).Observe(float64(time))
+	funcReapTime.WithLabelValues(funcName, funcAddress).Observe(time)
 }
 
 // IdleTime is the amount of time it took Reaper to find out the pod was idle
